🌼 separate different logic on each OS for child process to watch if master process exits
