--- prefork_dummy.go
@@ -1,13 +1,0 @@
-// +build !windows
-
-package fiber
-
-import (
-	"os/exec"
-)
-
-var dummyChildCmd = "go"
-
-func dummyCmd() *exec.Cmd {
-	return exec.Command(dummyChildCmd, "version")
-}
