--- prefork_dummy_windows.go
@@ -1,11 +1,0 @@
-package fiber
-
-import (
-	"os/exec"
-)
-
-var dummyChildCmd = "go"
-
-func dummyCmd() *exec.Cmd {
-	return exec.Command("cmd", "/C", dummyChildCmd, "version")
-}
